+++ src/System.IO.Pipelines/ReadableBuffer.cs
@@ -346,67 +346,9 @@ namespace System.IO.Pipelines
             var sb = new StringBuilder();
             foreach (var memory in this)
             {
-                int length = memory.Span.Length;
-                for (int i = 0; i < length; i++)
-                {
-                    char c = (char)memory.Span[i];
-                    AppendCharLiteral(c, sb);
-                }
+                SpanExtensions.AppendAsLiteral(memory.Span, sb);
             }
             return sb.ToString();
-        }
-
-        internal static void AppendCharLiteral(char c, StringBuilder sb)
-        {
-            switch (c)
-            {
-                case '\'':
-                    sb.Append(@"\'");
-                    break;
-                case '\"':
-                    sb.Append("\\\"");
-                    break;
-                case '\\':
-                    sb.Append(@"\\");
-                    break;
-                case '\0':
-                    sb.Append(@"\0");
-                    break;
-                case '\a':
-                    sb.Append(@"\a");
-                    break;
-                case '\b':
-                    sb.Append(@"\b");
-                    break;
-                case '\f':
-                    sb.Append(@"\f");
-                    break;
-                case '\n':
-                    sb.Append(@"\n");
-                    break;
-                case '\r':
-                    sb.Append(@"\r");
-                    break;
-                case '\t':
-                    sb.Append(@"\t");
-                    break;
-                case '\v':
-                    sb.Append(@"\v");
-                    break;
-                default:
-                    // ASCII printable character
-                    if (!char.IsControl(c))
-                    {
-                        sb.Append(c);
-                        // As UTF16 escaped character
-                    }
-                    else
-                    {
-                        sb.Append(@"\u");
-                        sb.Append(((int) c).ToString("x4"));
-                    }
-                    break;
-            }
         }
 
         /// <summary>
